Added per-computer network address support
Individual computers may now supply their network address during startup, and this address will be used when the server or user contacts the computer. If not supplied (for example, by older versions of the computer startup script), the system will fall back to concatenating the computer name and the Active Directory DNS domain to form the network address.