Disallow AD domain prefixes in usernames during login
The presence of an AD domain prefix causes generated RDP files to have an invalid username and user checking on lab machines to not work correctly.

This change does not reduce the functionality of the system, as the impacted components do not currently support users outside the default AD domain.